requirements.txt
numpy==1.21.5
matplotlib==3.5.2
seaborn==0.11.2
scikit-learn==1.0.2
tqdm==4.64.1
xgboost
xlsxwriter==3.0.3
pyyaml==6.0
scipy==1.9.1
torch
tensorflow
